# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/auth/password": {
      "patch": {
        "description": "任何已登录管理员都能改**自己**的密码，需要带上原密码（光有令牌不该能改口令）。\n\n改完直接返回**新的令牌**：库里令牌版本号已 +1，此前签发的令牌全部作废 —— 管理端应当用返回的 token 覆盖本地那一份，否则会被自己的系统踢回登录页。\n\n---\n\n**鉴权**：管理端令牌（`admin`）；本接口只操作当前登录管理员自身，不需要额外权限点。",
        "operationId": "adminAuth.changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAdminPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "改密成功（含新令牌）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminTokenDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "修改自己的密码",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminTokenDto"
        },
        "x-olua-permissions": []
      }
    },
...
    "/admin/accounts/{id}/password": {
      "patch": {
        "description": "玩家忘记密码时用这条（客户端没有找回流程）。\n\n**副作用**：该账号此前签发的全部令牌立即作废（服务端令牌版本号 +1）—— 账号被盗后改密码才能真正把对方踢下线，否则他手上那个 7 天有效期的令牌照样能用。\n\n请求体里的密码不会进操作日志（敏感字段自动打码为 `***`）。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`account:password`（重置玩家密码；角色：超级管理员 / 管理员）",
        "operationId": "adminAccount.resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "账号 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetAccountPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重置结果（不回显密码）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ResetPasswordResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "重置玩家账号的密码",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/ResetPasswordResultDto"
        },
        "x-olua-permissions": [
          "account:password"
        ]
      }
    },
...
    "/admin/admins": {
      "get": {
        "description": "keyword 模糊匹配管理员账号名；role 可按角色筛选。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`admin:read`（查看管理员；角色：超级管理员 / 管理员）",
        "operationId": "adminAdmin.list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "每页条数（1~100）",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字（不同接口含义见其说明）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "角色筛选",
            "schema": {
              "type": "string",
              "enum": [
                "super_admin",
                "admin",
                "viewer"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页结果（list/total/page/size）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminPageDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "管理员列表",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminPageDto"
        },
        "x-olua-permissions": [
          "admin:read"
        ],
        "x-olua-query-schema": {
          "$ref": "#/components/schemas/AdminQueryDto"
        }
      }
    },
    "/admin/admins/{id}": {
      "patch": {
        "description": "改角色或启停。允许改自己（超管轮值），但**最后一个启用中的超级管理员**不能被降级 / 停用 —— 避免没人能进后台。改完立即生效。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`admin:manage`（管理管理员；角色：超级管理员）",
        "operationId": "adminAdmin.update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "管理员 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改后的管理员",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "修改管理员",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminDto"
        },
        "x-olua-permissions": [
          "admin:manage"
        ]
      },
      "delete": {
        "description": "不可恢复；同样保护最后一个启用中的超级管理员。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`admin:manage`（管理管理员；角色：超级管理员）",
        "operationId": "adminAdmin.remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "管理员 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功（data 为 null）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true,
                          "additionalProperties": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "删除管理员",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": true,
        "x-olua-permissions": [
          "admin:manage"
        ]
      }
    },
    "/admin/admins/{id}/password": {
      "patch": {
        "description": "只有超级管理员能做（`admin:manage`）—— 管理员之间不能互相改密码。\n\n**副作用**：对方此前签发的令牌立即作废，他会被踢回登录页，需要用新密码重新登录。不触发「最后一个超管」保护：改密码不会让人失去登录能力，改的也不是身份与状态。管理员想改自己的密码应当走 `PATCH /admin/auth/password`（需要原密码）。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`admin:manage`（管理管理员；角色：超级管理员）",
        "operationId": "adminAdmin.resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "管理员 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetAdminPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重置结果（不回显密码）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ResetPasswordResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "重置某个管理员的密码",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/ResetPasswordResultDto"
        },
        "x-olua-permissions": [
          "admin:manage"
        ]
      }
    }
  },
  "info": {
    "title": "olua 游戏服务端 API",
    "description": "olua 客户端与管理端共用的服务端接口。\n\n**接口分三组**（左侧按组展示，顺序即权限层级）：\n1. **公共接口** —— 无需令牌（注册 / 登录 / 健康检查）\n2. **客户端** —— 需要 `player` 令牌（`/api/auth/login` 获取），只能操作自己账号的数据\n3. **管理端** —— 需要 `admin` 令牌（`/api/admin/auth/login` 获取），且每个接口还要求权限点\n\n**统一响应包裹**：所有接口（成功与失败）都是 `{ code, message, data, timestamp }`；\n`code = 0` 表示成功，其余见服务端 `common/constants/biz-code.ts`。\n\n**权限**：管理员分三级 —— 超级管理员 / 管理员 / 只读观察员，\n每个管理端接口的说明里都写了「所需权限」，由 `common/constants/permission.ts` 统一维护。\n请求头一律是 `Authorization: Bearer <token>`。\n\n---\n\n**这份文档是三端接口文件的唯一来源**：客户端 `assets/ui/utils/net/` 与管理端 `admin/src/api/`\n里的路径、类型、接口方法都由 `tools/gen-api.cjs` 从本文件生成，\n所以**改接口先改服务端**，再执行 `cd server && npm run gen:api` 重新生成（禁止手改生成物）。\n生成器读取的机器可读约定：\n- 每个接口的 `operationId` 形如 `<模块>.<方法>`（如 `role.save`），决定生成到哪个模块的哪个方法；\n- `x-olua-audience` 标明调用方（public / player / admin）；\n- `x-olua-permissions` 是管理端接口要求的权限点；\n- `x-olua-data-schema` 是响应里 `data` 的真实类型（没有该扩展 = 该接口无业务数据）。",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "player": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "客户端令牌（/api/auth/login 获取）"
      },
      "admin": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "管理端令牌（/api/admin/auth/login 获取）"
      }
    },
    "schemas": {
...
      "ChangeAdminPasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "当前密码"
          },
          "newPassword": {
            "type": "string",
            "description": "新密码（6~32 位）"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
...
      "ResetAccountPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "新密码（6~32 位）",
            "example": "new-pass-123"
          }
        },
        "required": [
          "password"
        ]
      },
...
      "AdminUpdateDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "目标角色",
            "enum": [
              "super_admin",
              "admin",
              "viewer"
            ]
          },
          "status": {
            "type": "string",
            "description": "目标状态",
            "enum": [
              "active",
              "disabled"
            ]
          }
        }
      },
      "ResetAdminPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "新密码（6~32 位）",
            "example": "new-pass-123"
          }
        },
        "required": [
          "password"
        ]
      },
      "ApiEnvelopeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "业务码，0 = 成功",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "ok"
          },
          "data": {
            "type": "object",
            "description": "业务数据（无数据时为 null）",
            "additionalProperties": true,
            "nullable": true
          },
          "timestamp": {
            "type": "number",
            "description": "服务端时间戳（毫秒）",
            "example": 1760000000000
          },
          "path": {
            "type": "string",
            "description": "出错时的请求路径",
            "example": "/api/auth/login"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
...
      "AdminDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "管理员 id"
          },
          "username": {
            "type": "string",
            "description": "管理员账号名"
          },
          "status": {
            "type": "string",
            "description": "状态：active 正常 / disabled 停用"
          },
          "role": {
            "type": "string",
            "description": "管理员角色：super_admin 超级管理员 / admin 管理员 / viewer 只读观察员",
            "enum": [
              "super_admin",
              "admin",
              "viewer"
            ]
          },
          "roleLabel": {
            "type": "string",
            "description": "角色中文名（界面直接展示）"
          },
          "permissions": {
            "type": "array",
            "description": "该角色拥有的权限点（管理端据此显示 / 隐藏按钮；服务端仍会独立校验）",
            "items": {
              "type": "string"
            },
            "example": [
              "stats:read",
              "account:read"
            ]
          },
          "createdAt": {
            "type": "number",
            "description": "创建时间（毫秒）"
          },
          "lastLoginAt": {
            "type": "number",
            "description": "最近登录时间（毫秒，从未登录为 null）",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "status",
          "role",
          "roleLabel",
          "permissions",
          "createdAt",
          "lastLoginAt"
        ]
      },
      "AdminTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "访问令牌（放 Authorization: Bearer <token>）"
          },
          "expiresIn": {
            "type": "string",
            "description": "令牌有效期",
            "example": "7d"
          },
          "admin": {
            "description": "管理员信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/AdminDto"
              }
            ]
          }
        },
        "required": [
          "token",
          "expiresIn",
          "admin"
        ]
      },
...
      "AdminPageDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "总条数",
            "example": 42
          },
          "page": {
            "type": "number",
            "description": "当前页码（从 1 开始）",
            "example": 1
          },
          "size": {
            "type": "number",
            "description": "每页条数",
            "example": 20
          },
          "list": {
            "description": "当前页的管理员列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "size",
          "list"
        ]
      },
...
      "ResetPasswordResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "被重置的对象 id"
          },
          "name": {
            "type": "string",
            "description": "被重置的对象名（账号名 / 管理员名，界面直接展示）"
          },
          "revokedTokens": {
            "type": "boolean",
            "description": "提示：口令变更后，该对象此前签发的令牌**已全部作废**，需要用它重新登录",
            "example": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "AdminQueryDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码，从 1 开始",
            "example": 1,
            "default": 1,
            "minimum": 1
          },
          "size": {
            "type": "number",
            "description": "每页条数（1~100）",
            "example": 20,
            "default": 20,
            "minimum": 1,
            "maximum": 100
          },
          "keyword": {
            "type": "string",
            "description": "关键字（不同接口含义见其说明）"
          },
          "role": {
            "type": "string",
            "description": "角色筛选",
            "enum": [
              "super_admin",
              "admin",
              "viewer"
            ]
          }
        }
      },
...
    }
  }
}